Safari-611.1.10/LayoutTests/js/dfg-value-to-int32-with-side-effect-expected.txt

Tests that a side-effecting ValueToInt32 only executes once when there is an OSR exit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS result.length is 2
PASS result[0] is 5
PASS counter is 100
PASS successfullyParsed is true

TEST COMPLETE
